notif [nfc]: Add Android-only asserts in NotificationDisplayManager methods
NotificationDisplayManager's public methods only get invoked on
Android, and their implementations basically assume the platform is
Android. So, add these asserts, to make that assumption clearer and
to fail early if we accidentally call one of these on iOS.

When we eventually take more hands-on control of notifications on
iOS, we might either adapt this class to support iOS too, or make a
new class for iOS and make this one's name and interface clearer
that it's Android-only.

Discussion:
  https://chat.zulip.org/#narrow/channel/243-mobile-team/topic/NotificationDisplayManager.20Android-only/near/2099196